Moved dataone publishing code into separate modules
