+++ gwvolman/tasks.py
@@ -24,7 +24,9 @@ from .utils import \
     new_user, _safe_mkdir, _get_api_key, \
     _get_container_config, _launch_container, _get_user_and_instance, \
     _build_image, DEPLOYMENT
-from .publish import publish_tale
+
+from .lib.dataone.publish import publish_tale
+
 from .constants import GIRDER_API_URL, InstanceStatus, ENABLE_WORKSPACES, \
     DEFAULT_USER, DEFAULT_GROUP, MOUNTPOINTS
 
@@ -382,33 +384,29 @@ def build_tale_image(self, tale_id):
         'last_build': build_time
     }
 
-
 @girder_job(title='Publish Tale')
 @app.task(bind=True)
 def publish(self,
             tale,
             dataone_node,
             dataone_auth_token,
-            girder_token,
-            userId):
+            user_id):
     """
     :param tale: The tale id
     :param dataone_node: The DataONE member node endpoint
     :param dataone_auth_token: The user's DataONE JWT
-    :param girder_token: The user's girder token
-    :param userId: The user's ID
+    :param user_id: The user's ID
     :type tale: str
     :type dataone_node: str
     :type dataone_auth_token: str
-    :type girder_token: str
-    :type userId: str
+    :type user_id: str
     """
-    publish_tale(self.job_manager,
+    publish_tale(self.girder_client,
+                 self.job_manager,
                  tale,
                  dataone_node,
                  dataone_auth_token,
-                 girder_token,
-                 userId)
+                 user_id)
 
 
 @girder_job(title='Import Tale')
